requirements.txt
flask==3.0.2
pymongo==4.6.2
python-dotenv==1.0.1
flask-jwt-extended==4.6.0
bcrypt==4.1.2
python-jose==3.3.0
email-validator==2.1.1
flask-cors==4.0.0 